refactor: move issue service to internal/issue
