+++ internal/attachment/service.go
@@ -42,6 +42,42 @@ func RemoveAttachment(ctx context.Context, m *core.Method, spath string) (*model
 }
 
 // ──────────────────────────────────────────────────────────────
+//  IssueAttachmentService
+// ──────────────────────────────────────────────────────────────
+
+// IssueAttachmentService handles communication with the issue attachment-related methods of the Backlog API.
+type IssueAttachmentService struct {
+	method *core.Method
+}
+
+// List returns a list of all attachments in the issue.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-issue-attachments
+func (s *IssueAttachmentService) List(ctx context.Context, issueIDOrKey string) ([]*model.Attachment, error) {
+	if err := validate.ValidateIssueIDOrKey(issueIDOrKey); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("issues", issueIDOrKey, "attachments")
+	return ListAttachments(ctx, s.method, spath)
+}
+
+// Remove removes a file attached to the issue.
+//
+// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/delete-issue-attachment
+func (s *IssueAttachmentService) Remove(ctx context.Context, issueIDOrKey string, attachmentID int) (*model.Attachment, error) {
+	if err := validate.ValidateIssueIDOrKey(issueIDOrKey); err != nil {
+		return nil, err
+	}
+	if err := validate.ValidateAttachmentID(attachmentID); err != nil {
+		return nil, err
+	}
+
+	spath := path.Join("issues", issueIDOrKey, "attachments", strconv.Itoa(attachmentID))
+	return RemoveAttachment(ctx, s.method, spath)
+}
+
+// ──────────────────────────────────────────────────────────────
 //  SpaceAttachmentService
 // ──────────────────────────────────────────────────────────────
 
@@ -141,6 +177,12 @@ func (s *WikiAttachmentService) Remove(ctx context.Context, wikiID, attachmentID
 // ──────────────────────────────────────────────────────────────
 //  Constructors
 // ──────────────────────────────────────────────────────────────
+
+func NewIssueAttachmentService(method *core.Method) *IssueAttachmentService {
+	return &IssueAttachmentService{
+		method: method,
+	}
+}
 
 func NewSpaceAttachmentService(method *core.Method) *SpaceAttachmentService {
 	return &SpaceAttachmentService{
