+++ service.go
@@ -1,9 +1,12 @@
 package backlog
 
 import (
+	"github.com/nattokin/go-backlog/internal/issue"
 	"github.com/nattokin/go-backlog/internal/space"
 	"github.com/nattokin/go-backlog/internal/wiki"
 )
+
+type IssueService = issue.IssueService
 
 type SpaceService = space.SpaceService
